turn off the forcesig option
Section 3.1.1. of the GPG manual says:

"Signature PIN

When set to "forced", gpg requests the entry of a PIN for each signature
operation. When set to "non forced", gpg may cache the PIN as long as
the card has not been removed from the reader."

Source: https://www.gnupg.org/howtos/card-howto/en/ch03.html

We should not need to force this Signature PIN when the card is on a
trusted computer belonging to the trusted user.

If the card is used on an untrusted computer by an untrusted user, it
SHOULD NOT be possible to sign without knowing the PIN.